Tested BallTracking. Successful. Extended Notes:
At the moment, when we have three or more balls, we disable our
collection, until we trip the exit switch enough times that
we have fewer than 3 balls.
There are some problems, namely that we cannot repel balls in
collection disabled mode as it is written right now, but what
we were effectively looking for is working.